Change loop index type from unsigned int to size_t

+++ src/goto-symex/witnesses.cpp
@@ -1134,7 +1134,7 @@ bool find_nondet_in_expr(const expr2tc &expr)
     return is_nondet_symbol_name(to_symbol2t(expr).thename.as_string());
 
   unsigned int n = expr->get_num_sub_exprs();
-  for (unsigned int i = 0; i < n; ++i)
+  for (size_t i = 0; i < n; ++i)
   {
     const expr2tc *sub = expr->get_sub_expr(i);
     if (sub && find_nondet_in_expr(*sub))
